requirements.txt
anthropic>=0.49.0
streamlit>=1.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pypdf>=5.0.0
sqlalchemy>=2.0.0
folium>=0.18.0
streamlit-folium>=0.22.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
feedparser>=6.0.11
rapidfuzz>=3.0.0
pandas>=2.0.0
